+++ apps/admin/package.json
@@ -22,17 +22,16 @@
     "react-dom": "catalog:",
     "slugify": "1.6.9",
     "sonner": "2.0.7",
-    "workflow": "catalog:",
     "zod": "4.4.3"
   },
   "description": "Admin application for content management",
   "devDependencies": {
     "@tailwindcss/postcss": "4.3.0",
+    "@vercel/config": "0.5.1",
     "@tailwindcss/typography": "0.5.20",
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
-    "@vercel/config": "0.5.1",
     "@ykzts/tsconfig": "workspace:*",
     "postcss": "8.5.15",
     "shadcn": "4.11.0",
