+++ apps/blog/app/api/blog/cron/publish/route.ts
@@ -1,8 +1,7 @@
+import { revalidateTag } from "next/cache";
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { start } from "workflow/api";
 import { supabaseAdmin } from "@/lib/supabase/client";
-import { publishScheduledPosts } from "@/workflows/publish-scheduled-posts";
 
 export async function GET(request: NextRequest) {
   // Verify Vercel Cron Secret
@@ -21,12 +20,47 @@ export async function GET(request: NextRequest) {
   }
 
   try {
-    // Fire-and-forget: enqueue the durable workflow.
-    // The cron handler returns immediately; do not await run.returnValue.
-    const run = await start(publishScheduledPosts, []);
+    // Find scheduled posts that should be published now
+    const now = new Date().toISOString();
+    const { data: scheduledPosts, error: queryError } = await supabaseAdmin
+      .from("posts")
+      .select("id, slug, title")
+      .eq("status", "scheduled")
+      .lte("published_at", now);
+
+    if (queryError) {
+      throw queryError;
+    }
+
+    if (!scheduledPosts || scheduledPosts.length === 0) {
+      return NextResponse.json({
+        message: "No posts to publish",
+        publishedCount: 0,
+      });
+    }
+
+    // Update posts to published status
+    const postIds = scheduledPosts.map((post) => post.id);
+    const { error: updateError } = await supabaseAdmin
+      .from("posts")
+      .update({ status: "published" })
+      .in("id", postIds);
+
+    if (updateError) {
+      throw updateError;
+    }
+
+    // Invalidate cache
+    revalidateTag("posts", "max");
+
     return NextResponse.json({
-      message: "Workflow started for publishing scheduled posts",
-      runId: run.runId,
+      message: `Published ${scheduledPosts.length} post(s)`,
+      posts: scheduledPosts.map((post) => ({
+        id: post.id,
+        slug: post.slug,
+        title: post.title,
+      })),
+      publishedCount: scheduledPosts.length,
     });
   } catch (error) {
     console.error("Error publishing scheduled posts:", error);
