+++ apps/blog/next.config.ts
@@ -1,7 +1,7 @@
 import { withMicrofrontends } from "@vercel/microfrontends/next/config";
 import { getSupabaseImageConfig } from "@ykzts/supabase/next-image-config";
+
 import type { NextConfig } from "next";
-import { withWorkflow } from "workflow/next";
 
 const nextConfig: NextConfig = {
   cacheComponents: true,
@@ -33,4 +33,4 @@ const nextConfig: NextConfig = {
   typedRoutes: true,
 };
 
-export default withWorkflow(withMicrofrontends(nextConfig));
+export default withMicrofrontends(nextConfig);
