+++ apps/blog/package.json
@@ -18,15 +18,15 @@
     "next-themes": "0.4.6",
     "react": "catalog:",
     "react-dom": "catalog:",
-    "server-only": "catalog:",
+    "server-only": "0.0.1",
     "shiki": "4.2.0",
     "tailwind-merge": "3.6.0",
-    "workflow": "catalog:",
     "zod": "4.4.3"
   },
   "description": "Next.js blog application for ykzts.blog",
   "devDependencies": {
     "@portabletext/types": "4.0.2",
+    "@vercel/config": "0.5.1",
     "@tailwindcss/postcss": "4.3.0",
     "@tailwindcss/typography": "0.5.20",
     "@testing-library/dom": "10.4.1",
@@ -36,7 +36,6 @@
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
-    "@vercel/config": "0.5.1",
     "@vitejs/plugin-react": "6.0.2",
     "@ykzts/tsconfig": "workspace:*",
     "jsdom": "29.1.1",
