+++ apps/memo/package.json
@@ -12,7 +12,7 @@
     "next": "catalog:",
     "react": "catalog:",
     "react-dom": "catalog:",
-    "server-only": "catalog:",
+    "server-only": "0.0.1",
     "sonner": "2.0.7"
   },
   "description": "Memo application for ykzts",
