+++ apps/portfolio/app/_actions/contact.ts
@@ -1,10 +1,12 @@
 "use server";
 
+import { getSiteName } from "@ykzts/site-config";
 import { getProfile } from "@ykzts/supabase/queries";
 import { checkBotId } from "botid/server";
-import { start } from "workflow/api";
+import { Resend } from "resend";
 import { z } from "zod";
-import { sendContactEmail } from "@/workflows/send-contact-email";
+
+const resend = new Resend(process.env.RESEND_API_KEY);
 
 const contactFormSchema = z.object({
   email: z.email("有効なメールアドレスを入力してください"),
@@ -68,9 +70,44 @@ export async function submitContactForm(
     // Validate form data
     const validatedData = contactFormSchema.parse(data);
 
-    // Fire-and-forget: enqueue the durable email workflow.
-    // No need to await run.returnValue here — the HTTP handler should not block on queued work.
-    await start(sendContactEmail, [{ contactEmail, data: validatedData }]);
+    // Send email using Resend
+    if (!(process.env.RESEND_API_KEY && contactEmail)) {
+      if (!process.env.RESEND_API_KEY) {
+        console.error("RESEND_API_KEY is not configured.");
+      }
+
+      if (!contactEmail) {
+        console.error("contactEmail in profile is not configured.");
+      }
+
+      return {
+        error:
+          "メール送信の設定が正しくありません。管理者にお問い合わせください。",
+        formData: data,
+        success: false,
+      };
+    }
+
+    const siteName = getSiteName();
+    const fromAddress = process.env.MAIL_FROM_ADDRESS ?? "no-reply@example.com";
+
+    const emailResult = await resend.emails.send({
+      from: `${validatedData.name} via ${siteName} <${fromAddress}>`,
+      replyTo: `${validatedData.name} <${validatedData.email}>`,
+      subject: `[お問い合わせ] ${validatedData.subject}`,
+      text: validatedData.message,
+      to: [contactEmail],
+    });
+
+    if (emailResult.error) {
+      console.error("Resend error:", emailResult.error);
+      return {
+        error:
+          "メールの送信に失敗しました。しばらくしてからもう一度お試しください。",
+        formData: data,
+        success: false,
+      };
+    }
 
     return {
       success: true,
