+++ apps/portfolio/package.json
@@ -25,7 +25,6 @@
     "shiki": "4.2.0",
     "sonner": "2.0.7",
     "tailwind-merge": "3.6.0",
-    "workflow": "catalog:",
     "zod": "4.4.3"
   },
   "description": "Next.js portfolio website showcasing projects, skills, and professional experience",
