+++ pnpm-workspace.yaml
@@ -12,16 +12,12 @@ catalog:
   next: 16.2.7
   react: 19.2.7
   react-dom: 19.2.7
-  server-only: 0.0.1
   typescript: 6.0.3
   vitest: 4.1.8
-  workflow: 4.4.0
 
 allowBuilds:
-  '@swc/core': true
   '@tailwindcss/oxide': true
   better-sqlite3: false
-  cbor-extract: true
   core-js-pure: false
   dtrace-provider: false
   esbuild: true
